Improve openebs-provisioner code structure.
This make openebs-provisioner code structure modular and more
readable.
