+++ openebs/openebs-provisioner.go
@@ -18,12 +18,8 @@ limitations under the License.
 package main
 
 import (
-	"bytes"
-	"encoding/json"
 	"errors"
 	"flag"
-	"io/ioutil"
-	"net/http"
 	"os"
 	"time"
 
@@ -31,7 +27,8 @@ import (
 
 	"github.com/golang/glog"
 	"github.com/kubernetes-incubator/external-storage/lib/controller"
-	yaml "gopkg.in/yaml.v2"
+	mApiv1 "github.com/kubernetes-incubator/external-storage/openebs/pkg/v1"
+	mayav1 "github.com/kubernetes-incubator/external-storage/openebs/types/v1"
 	"k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/wait"
@@ -51,43 +48,6 @@ const (
 	termLimit                 = controller.DefaultTermLimit
 )
 
-//VsmSpec holds the config for creating a VSM
-type VsmSpec struct {
-	Kind       string `yaml:"kind"`
-	APIVersion string `yaml:"apiVersion"`
-	Metadata   struct {
-		Name   string `yaml:"name"`
-		Labels struct {
-			Storage string `yaml:"volumeprovisioner.mapi.openebs.io/storage-size"`
-		}
-	} `yaml:"metadata"`
-}
-
-// Volume is a command implementation struct
-type Volume struct {
-	Spec struct {
-		AccessModes interface{} `json:"AccessModes"`
-		Capacity    interface{} `json:"Capacity"`
-		ClaimRef    interface{} `json:"ClaimRef"`
-		OpenEBS     struct {
-			VolumeID string `json:"volumeID"`
-		} `json:"OpenEBS"`
-		PersistentVolumeReclaimPolicy string `json:"PersistentVolumeReclaimPolicy"`
-		StorageClassName              string `json:"StorageClassName"`
-	} `json:"Spec"`
-
-	Status struct {
-		Message string `json:"Message"`
-		Phase   string `json:"Phase"`
-		Reason  string `json:"Reason"`
-	} `json:"Status"`
-	Metadata struct {
-		Annotations       interface{} `json:"annotations"`
-		CreationTimestamp interface{} `json:"creationTimestamp"`
-		Name              string      `json:"name"`
-	} `json:"metadata"`
-}
-
 type openEBSProvisioner struct {
 	// Maya-API Server URI running in the cluster
 	mapiURI string
@@ -105,7 +65,7 @@ func NewOpenEBSProvisioner(client kubernetes.Interface) controller.Provisioner {
 	}
 
 	//TODO - HandleError Cases
-	mayaServiceURI := "http://" + getMayaClusterIP(client) + ":5656"
+	mayaServiceURI := "http://" + mApiv1.GetMayaClusterIP(client) + ":5656"
 	os.Setenv("MAPI_ADDR", mayaServiceURI)
 
 	return &openEBSProvisioner{
@@ -118,19 +78,20 @@ var _ controller.Provisioner = &openEBSProvisioner{}
 
 // Provision creates a storage asset and returns a PV object representing it.
 func (p *openEBSProvisioner) Provision(options controller.VolumeOptions) (*v1.PersistentVolume, error) {
-	//path := "/var/openebs/" + options.PVName
 
-	//TODO - Issue a request to Maya API Server to create a volume
-	var volume Volume
+	//Issue a request to Maya API Server to create a volume
+	var volume mayav1.Volume
+
 	volSize := options.PVC.Spec.Resources.Requests[v1.ResourceName(v1.ResourceStorage)]
+
 	//TODO - Need to change the size as a value that Maya Server can accept
-	err := createVsm(options.PVName, volSize.String())
+	err := mApiv1.CreateVsm(options.PVName, volSize.String())
 	if err != nil {
 		glog.Fatalf("Error creating volume: %v", err)
 		return nil, err
 	}
 
-	err = listVsm(options.PVName, &volume)
+	err = mApiv1.ListVsm(options.PVName, &volume)
 	if err != nil {
 		glog.Fatalf("Error getting volume details: %v", err)
 		return nil, err
@@ -191,141 +152,8 @@ func (p *openEBSProvisioner) Delete(volume *v1.PersistentVolume) error {
 	}
 
 	//TODO - Issue a delete request to Maya API Server
-	deleteVsm(volume.Name)
+	mApiv1.DeleteVsm(volume.Name)
 
-	return nil
-}
-
-func getMayaClusterIP(client kubernetes.Interface) string {
-	clusterIP := "127.0.0.1"
-
-	//Fetch the Maya ClusterIP using the Maya API Server Service
-	sc, err := client.CoreV1().Services("default").Get("maya-apiserver-service", metav1.GetOptions{})
-	if err != nil {
-		glog.Fatalf("Error getting maya-api-server IP Address: %v", err)
-	}
-
-	clusterIP = sc.Spec.ClusterIP
-	glog.Infof("Maya Cluster IP: %v", clusterIP)
-
-	return clusterIP
-}
-
-// createVsm to create the Vsm through a API call to m-apiserver
-func createVsm(vname string, size string) error {
-
-	var vs VsmSpec
-
-	addr := os.Getenv("MAPI_ADDR")
-	if addr == "" {
-		err := errors.New("MAPI_ADDR environment variable not set")
-		glog.Fatalf("Error getting maya-api-server IP Address: %v", err)
-		return err
-	}
-	url := addr + "/latest/volumes/"
-
-	vs.Kind = "PersistentVolumeClaim"
-	vs.APIVersion = "v1"
-	vs.Metadata.Name = vname
-	vs.Metadata.Labels.Storage = size
-
-	//Marshal serializes the value provided into a YAML document
-	yamlValue, _ := yaml.Marshal(vs)
-
-	glog.Infof("VSM Spec Created:\n%v\n", string(yamlValue))
-
-	req, err := http.NewRequest("POST", url, bytes.NewBuffer(yamlValue))
-
-	req.Header.Add("Content-Type", "application/yaml")
-
-	c := &http.Client{
-		Timeout: timeout,
-	}
-	resp, err := c.Do(req)
-	if err != nil {
-		glog.Fatalf("http.Do() error: : %v", err)
-		return err
-	}
-	defer resp.Body.Close()
-
-	data, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		glog.Fatalf("ioutil.ReadAll() error: : %v", err)
-		return err
-	}
-
-	code := resp.StatusCode
-	if code != http.StatusOK {
-		glog.Fatalf("Status error: %v\n", http.StatusText(code))
-		os.Exit(1)
-	}
-	glog.Infof("VSM Successfully Created:\n%v\n", string(data))
-	return nil
-}
-
-// listVsm to get the info of Vsm through a API call to m-apiserver
-func listVsm(vname string, obj interface{}) error {
-
-	addr := os.Getenv("MAPI_ADDR")
-	if addr == "" {
-		err := errors.New("MAPI_ADDR environment variable not set")
-		glog.Fatalf("Error getting maya-api-server IP Address: %v", err)
-		return err
-	}
-	url := addr + "/latest/volumes/info/" + vname
-
-	glog.Infof("Get details for VSM :%v", string(vname))
-
-	req, err := http.NewRequest("GET", url, nil)
-	c := &http.Client{
-		Timeout: timeout,
-	}
-	resp, err := c.Do(req)
-	if err != nil {
-		glog.Fatalf("http.Do() error: : %v", err)
-		return err
-	}
-	defer resp.Body.Close()
-
-	code := resp.StatusCode
-	if code != http.StatusOK {
-		glog.Fatalf("Status error: %v\n", http.StatusText(code))
-		os.Exit(1)
-	}
-	glog.Info("VSM Details Successfully Retrieved")
-	return json.NewDecoder(resp.Body).Decode(obj)
-}
-
-// deleteVsm to get delete Vsm through a API call to m-apiserver
-func deleteVsm(vname string) error {
-
-	addr := os.Getenv("MAPI_ADDR")
-	if addr == "" {
-		err := errors.New("MAPI_ADDR environment variable not set")
-		glog.Fatalf("Error getting maya-api-server IP Address: %v", err)
-		return err
-	}
-	url := addr + "/latest/volumes/delete/" + vname
-
-	glog.Infof("Delete VSM :%v", string(vname))
-
-	req, err := http.NewRequest("GET", url, nil)
-	c := &http.Client{
-		Timeout: timeout,
-	}
-	resp, err := c.Do(req)
-	if err != nil {
-		glog.Fatalf("http.Do() error: : %v", err)
-		return err
-	}
-	defer resp.Body.Close()
-
-	code := resp.StatusCode
-	if code != http.StatusOK {
-		glog.Fatalf("Status error: %v\n", http.StatusText(code))
-		os.Exit(1)
-	}
-	glog.Info("VSM Deleted Successfully initiated")
 	return nil
 }
 
@@ -364,20 +192,12 @@ func main() {
 		provisionerName,
 		openEBSProvisioner,
 		serverVersion.GitVersion)
-		// resyncPeriod,
-		// exponentialBackOffOnError,
-		// failedRetryThreshold,
-		// leasePeriod,
-		// renewDeadline,
-		// retryPeriod,
-		// termLimit)
+	// resyncPeriod,
+	// exponentialBackOffOnError,
+	// failedRetryThreshold,
+	// leasePeriod,
+	// renewDeadline,
+	// retryPeriod,
+	// termLimit)
 	pc.Run(wait.NeverStop)
 }
-
-/*
-	client kubernetes.Interface,
-	provisionerName string,
-	provisioner Provisioner,
-	kubeVersion string,
-	options ...func(*ProvisionController) error,
-*/
